# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shalomut Map API (מפת שלומות)",
    "version": "1.0.0",
    "description": "Clean Data Layer & REST API for Shalomut Map — visualizing school staff wellbeing across 8 organic dimensions. Implements ADR-001 (Raw Data & Privacy Threshold Control)."
  },
  "servers": [
    {
      "url": "https://shalomut-map-demo-ui-redesign.vercel.app",
      "description": "Staging Environment"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
    "/api/rounds": {
      "get": {
        "tags": ["Rounds"],
        "summary": "Fetch default survey round",
        "description": "Returns details of the default active survey round.",
        "security": [
          { "basicAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of round details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "round": {
                      "$ref": "#/components/schemas/SurveyRound"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Manager organization scope is required"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Rounds"],
        "summary": "Create new survey round",
        "description": "Generates a new survey round for an organization with a unique share code (SHALOM-XXXX).",
        "security": [
          { "basicAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoundInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Survey round created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "round": {
                      "$ref": "#/components/schemas/SurveyRound"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Organization not found in the manager scope"
          },
          "500": {
            "description": "Failed to create round",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Persistent storage is not configured for this deployment"
          }
        }
      }
    },
...
    "/api/rounds/{roundId}/trigger-ai": {
      "post": {
        "tags": ["AI Integration"],
        "summary": "Request AI processing for a survey round",
        "description": "Forwards a round_closed event to AI_SERVICE_URL without a dynamic callback target. The AI service derives its callback URL from its trusted DATA_LAYER_CALLBACK_URL configuration and the round ID. The endpoint fails closed when the AI service is unavailable.",
        "security": [
          { "basicAuth": [] }
        ],
        "parameters": [
          {
            "name": "roundId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "4f88af95-0ec4-4872-85ac-67466c3cd8f2"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "AI service accepted the event"
          },
          "404": {
            "description": "Survey round not found in the manager scope"
          },
          "502": {
            "description": "AI service returned a non-success response"
          },
          "503": {
            "description": "AI service could not be reached"
          },
          "500": {
            "description": "Failed to construct or dispatch the request"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Manager credential bound server-side to MANAGER_ORGANIZATION_ID."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Shared service secret. Required only when the corresponding environment variable is configured."
      }
    },
    "schemas": {
      "RoundStatus": {
        "type": "string",
        "enum": ["draft", "active", "closed", "archived"],
        "example": "active"
      },
...
      "WellbeingDimensionId": {
        "type": "string",
        "enum": [
          "self-expression",
          "professional-competence",
          "social-resource",
          "balance",
          "management-support",
          "certainty",
          "organizational-climate",
          "meaning"
        ],
        "example": "self-expression"
      },
      "CreateRoundInput": {
        "type": "object",
        "required": ["organizationId", "title"],
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid",
            "example": "c93f8832-8d0d-43cb-b8d2-cdd6e45683af"
          },
          "title": {
            "type": "string",
            "example": "סקר שלומות - מחצית א' תשפ\"ו"
          },
          "privacyThreshold": {
            "type": "integer",
            "default": 10,
            "example": 10
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "RoundBackgroundContext": {
        "type": "object",
        "required": [
          "notes",
          "audience",
          "sicknessDaysThisQuarter",
          "newStaffMembers",
          "studentCount",
          "socioEconomicIndex",
          "classesPerGrade"
        ],
        "properties": {
          "notes": { "type": "string" },
          "audience": { "type": "string", "example": "כלל צוות ההוראה" },
          "sicknessDaysThisQuarter": { "type": "integer", "minimum": 0 },
          "newStaffMembers": { "type": "integer", "minimum": 0 },
          "studentCount": { "type": "integer", "minimum": 0 },
          "socioEconomicIndex": { "type": "integer", "minimum": 1, "maximum": 10 },
          "classesPerGrade": {
            "type": "object",
            "additionalProperties": { "type": "integer", "minimum": 0 }
          }
        }
      },
...
      "SurveyDefinitionQuestion": {
        "type": "object",
        "required": ["id", "text", "dimensionId", "required", "enabled", "answerMode"],
        "properties": {
          "id": { "type": "string", "example": "q1" },
          "text": { "type": "string" },
          "dimensionId": { "$ref": "#/components/schemas/WellbeingDimensionId" },
          "required": { "type": "boolean" },
          "enabled": { "type": "boolean" },
          "answerMode": { "type": "string", "example": "סקאלת צבעים" }
        }
      },
      "SurveyDefinition": {
        "type": "object",
        "required": [
          "title",
          "audience",
          "estimatedMinutes",
          "minimumResponses",
          "introText",
          "anonymityText",
          "questions"
        ],
        "properties": {
          "title": { "type": "string" },
          "audience": { "type": "string" },
          "estimatedMinutes": { "type": "number", "minimum": 1 },
          "minimumResponses": { "type": "integer", "minimum": 10 },
          "introText": { "type": "string" },
          "anonymityText": { "type": "string" },
          "questions": {
            "type": "array",
            "minItems": 24,
            "items": { "$ref": "#/components/schemas/SurveyDefinitionQuestion" }
          }
        }
      },
      "SurveyRound": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "example": "4f88af95-0ec4-4872-85ac-67466c3cd8f2" },
          "organizationId": { "type": "string", "format": "uuid", "example": "c93f8832-8d0d-43cb-b8d2-cdd6e45683af" },
          "title": { "type": "string", "example": "סקר שלומות - מחצית א'" },
          "status": { "$ref": "#/components/schemas/RoundStatus" },
          "shareCode": { "type": "string", "example": "SHALOM-7K2M" },
          "privacyThreshold": { "type": "integer", "example": 10 },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "backgroundContext": { "$ref": "#/components/schemas/RoundBackgroundContext" },
          "surveyDefinition": { "$ref": "#/components/schemas/SurveyDefinition" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Failed to process request" }
        }
      }
    }
  }
}